Added RenderingTarget
The idea behind this protocol is that the event controller
polyfill will not actually be able to switch between player
instances without having access to where the player is
rendering to. So this will have to be provided by the consumer.
Example implementations could be switching player on
AVPlayerViewController, or switching player on an AVPlayerLayer.