Restrict date/time mappings when not likely to keep up
When seeking it can be the case that many times the effective rate
changes to 0, but the currentDate() is not reliable, and this seems
associated with when the player is not likely to keep up. Based on
this we make a further restriction to date/time mapping such that
it is only recorded when playback is likely to keep up.